refactor: move commands files
